rename: (AWSConfig) config 패키지로 이동 / security 패키지도 이동
