+++ src/main/java/com/kl/grooveo/base/config/security/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package com.kl.grooveo.base.security;
+package com.kl.grooveo.base.config.security;
 
 import java.util.Collection;
 import java.util.Map;
